+++ src/Opulence/Ioc/Tests/Bootstrappers/Inspection/Mocks/Bar.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2019 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+
+declare(strict_types=1);
+
+namespace Opulence\Ioc\Tests\Bootstrappers\Inspection\Mocks;
+
+/**
+ * Defines a fake concrete class for use in testing
+ */
+final class Bar implements IBar
+{
+    // Don't do anything
+}